Module

Feature selection has been one of the important steps in machine learning. Some of the advantages of feature selection are the performance increase and may prevent model to overfits.

This technique selects features based on their importance, which is defined by how much impact does the feature have to the target variable. This package helps selecting the important features based on the correlation and p-value.

Selectors

Correlation based feature selector

FeatureSelector.CorrelationBasedFeatureSelector — Type

CorrelationBasedFeatureSelector has the following fields:

k::Int64 - Select top k features with the highest correlation to target variable. You could ignore this by specifying k <= 0. This defaults to 0.

threshold::Float64 - Select features with correlation more than or equal to threshold. To ignore, simply set threshold to 0 (default behavior).

P-value based feature selector

FeatureSelector.PValueBasedFeatureSelector — Type

PValueBasedFeatureSelector has the following fields:

k::Int64 - Select top k features with the highest correlation to target variable. You could ignore this by specifying k <= 0. This defaults to 0.

threshold::Float64 - Select features with correlation less than or equal to threshold. Note that, in P-value, a feature is considered important when the value is close to 0. To ignore, simply set threshold to 0 (default behavior).

method::Symbol - Type of test to get the p-value. The available option are :FTest and :ChiSq. :ChiSq expects the input and target are in non-negative integers.

Select feature function

FeatureSelector.select_features — Method

function select_features(selector,
                         X::DataFrame,
                         y::Vector;
                         verbose::Bool=false,
                         return_val::Bool=false)

Select features based on the importance, which is defined by selector to target y. if verbose is true, logs will be printed - this defaults to false. If return_val is true, this function will return only the feature feature names, otherwise, tuple of selected feature names and the correlation value are returned.

If you have feature X_data as matrix and feature names X_features as a Vector, you can replace X with X_data and X_features (in this order).

Example

julia> using RDatasets, FeatureSelector

julia> boston = dataset("MASS", "Boston");

julia> selector = CorrelationBasedFeatureSelector(k=5)
CorrelationBasedFeatureSelector(5, 0.0)

julia> select_features(selector, boston[:, Not(:MedV)], boston.MedV)
5-element Array{Symbol,1}:
 :LStat
 :Rm
 :PTRatio
 :Indus
 :Tax

Other util functions

FeatureSelector.one_hot_encode — Method

function one_hot_encode(df::DataFrame;
                        cols::Vector{Symbol}=Vector{Symbol}(),
                        drop_original::Bool=false)

Utility function to perform one-hot-encoding in DataFrame. This will add new columns with names <original_col_name>_<value>.

Following options can be passed to modify behavior.

cols - Vector of Symbol to specify which columns to be encoded. Defaults to empty, which means all features will be encoded.

drop_original - If true, this will drop the original feature set from resulting DataFrame. This defaults to false.

Example

julia> using RDatasets, FeatureSelector

julia> titanic = dataset("datasets", "Titanic");

julia> first(one_hot_encode(titanic[:, [:Class, :Sex, :Age]]), 3)
3×11 DataFrames.DataFrame. Omitted printing of 5 columns
│ Row │ Class  │ Sex    │ Age    │ Class_1st │ Class_2nd │ Class_3rd │
│     │ String │ String │ String │ Bool      │ Bool      │ Bool      │
├─────┼────────┼────────┼────────┼───────────┼───────────┼───────────┤
│ 1   │ 1st    │ Male   │ Child  │ 1         │ 0         │ 0         │
│ 2   │ 2nd    │ Male   │ Child  │ 0         │ 1         │ 0         │
│ 3   │ 3rd    │ Male   │ Child  │ 0         │ 0         │ 1         │

julia> first(one_hot_encode(titanic[:, [:Class, :Sex, :Age]], cols=[:Class], drop_original=true), 3)
3×6 DataFrames.DataFrame
│ Row │ Sex    │ Age    │ Class_1st │ Class_2nd │ Class_3rd │ Class_Crew │
│     │ String │ String │ Bool      │ Bool      │ Bool      │ Bool       │
├─────┼────────┼────────┼───────────┼───────────┼───────────┼────────────┤
│ 1   │ Male   │ Child  │ 1         │ 0         │ 0         │ 0          │
│ 2   │ Male   │ Child  │ 0         │ 1         │ 0         │ 0          │
│ 3   │ Male   │ Child  │ 0         │ 0         │ 1         │ 0          │